+++ WTF/Account/NAME/SavedVariables/CooldownManagerCentered.lua
@@ -1,0 +1,16 @@
+
+CooldownManagerCenteredDB = {
+["profileKeys"] = {
+["Pulcinhell - Draenor"] = "Default",
+["Formaggino - Draenor"] = "Default",
+["Nutellone - Draenor"] = "Default",
+["Panzarotto - Draenor"] = "Default",
+},
+["profiles"] = {
+["Default"] = {
+["cooldownManager_centerUtility"] = true,
+["cooldownManager_centerBuffIcons"] = true,
+["cooldownManager_alignBuffBars"] = true,
+},
+},
+}