+++ WTF/Account/NAME/SavedVariables/Masque.lua
@@ -1,0 +1,164 @@
+
+MasqueDB = {
+["namespaces"] = {
+},
+["profileKeys"] = {
+["Pulcinhell - Draenor"] = "Default",
+["Formaggino - Draenor"] = "Default",
+["Nutellone - Draenor"] = "Default",
+["Panzarotto - Draenor"] = "Default",
+},
+["profiles"] = {
+["Default"] = {
+["Groups"] = {
+["Blizzard Inventory_MailFrame"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Action Bars_StanceBar"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Inventory_LootFrame"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Inventory_PaperDollFrame"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Buffs_Debuffs"] = {
+["Inherit"] = false,
+},
+["Blizzard Action Bars_SpellFlyout"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Action Bars_MultiBarBottomRight"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Buffs"] = {
+["Inherit"] = false,
+},
+["Blizzard Inventory_ContainerFrames"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Inventory_ContainerFrameCombinedBags"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Action Bars_MultiBarLeft"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Action Bars"] = {
+["Inherit"] = false,
+},
+["Blizzard Inventory"] = {
+["Inherit"] = false,
+},
+["Blizzard Inventory_BankPanel"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Cooldown Manager Tweaks_Custom Tracker"] = {
+["Inherit"] = false,
+},
+["Blizzard Inventory_ContainerFrame1"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Action Bars_MultiBar6"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Action Bars_MultiBar7"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Inventory_MerchantFrame"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Action Bars_ActionBar"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Action Bars_PetBattleFrame"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Buffs_Buffs"] = {
+["Inherit"] = false,
+},
+["Blizzard Inventory_ContainerFrame6"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Action Bars_MultiBarBottomLeft"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Cooldown Manager Tweaks_Buff Tracker"] = {
+["Inherit"] = false,
+},
+["Blizzard Action Bars_OverrideActionBar"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Cooldown Manager Tweaks_Essential Cooldowns"] = {
+["Inherit"] = false,
+},
+["Blizzard Action Bars_PetBar"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Inventory_GuildBankFrame"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Cooldown Manager Tweaks_Utility Cooldowns"] = {
+["Inherit"] = false,
+},
+["Blizzard Action Bars_PossessBar"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Action Bars_MultiBar5"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Action Bars_CooldownViewer"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Action Bars_ExtraAbilityContainer"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Action Bars_MultiBarRight"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Blizzard Inventory_InspectPaperDollFrame"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+["Cooldown Manager Tweaks"] = {
+["Inherit"] = false,
+},
+["Blizzard Inventory_EquipmentFlyoutFrame"] = {
+["Upgraded"] = true,
+["Inherit"] = false,
+},
+},
+["LDB"] = {
+["position"] = 0,
+["hide"] = true,
+},
+["API_VERSION"] = 110207,
+},
+},
+}