+++ WTF/Account/NAME/SavedVariables/TomTom.lua
@@ -1,0 +1,91 @@
+
+TomTomDB = {
+["profileKeys"] = {
+["Panuozzo - Draenor"] = "Default",
+["Scaloppina - Draenor"] = "Default",
+["Gelatino - Draenor"] = "Default",
+["Formaggino - Draenor"] = "Default",
+["Sfilatino - Draenor"] = "Default",
+["Cirus - Argent Dawn"] = "Default",
+["Pulcinhell - Draenor"] = "Default",
+["Panuozzo - Argent Dawn"] = "Default",
+["Panzarotto - Draenor"] = "Default",
+["Giandora - Argent Dawn"] = "Default",
+["Percoca - Draenor"] = "Default",
+["Branthar - Argent Dawn"] = "Default",
+["Nutellone - Draenor"] = "Default",
+["Ossosacro - Draenor"] = "Default",
+},
+["profiles"] = {
+["Default"] = {
+["arrow"] = {
+["position"] = {
+"CENTER",
+nil,
+"CENTER",
+0,
+0,
+},
+},
+["block"] = {
+["enable"] = false,
+["position"] = {
+"CENTER",
+nil,
+"CENTER",
+52.30772399902344,
+-104.6151962280273,
+},
+},
+},
+},
+}
+TomTomWaypoints = nil
+TomTomWaypointsM = {
+["profileKeys"] = {
+["Panuozzo - Draenor"] = "Panuozzo - Draenor",
+["Scaloppina - Draenor"] = "Scaloppina - Draenor",
+["Gelatino - Draenor"] = "Gelatino - Draenor",
+["Formaggino - Draenor"] = "Formaggino - Draenor",
+["Sfilatino - Draenor"] = "Sfilatino - Draenor",
+["Cirus - Argent Dawn"] = "Cirus - Argent Dawn",
+["Pulcinhell - Draenor"] = "Pulcinhell - Draenor",
+["Panuozzo - Argent Dawn"] = "Panuozzo - Argent Dawn",
+["Panzarotto - Draenor"] = "Panzarotto - Draenor",
+["Giandora - Argent Dawn"] = "Giandora - Argent Dawn",
+["Percoca - Draenor"] = "Percoca - Draenor",
+["Branthar - Argent Dawn"] = "Branthar - Argent Dawn",
+["Nutellone - Draenor"] = "Nutellone - Draenor",
+["Ossosacro - Draenor"] = "Ossosacro - Draenor",
+},
+["profiles"] = {
+["Panuozzo - Draenor"] = {
+},
+["Scaloppina - Draenor"] = {
+},
+["Gelatino - Draenor"] = {
+},
+["Formaggino - Draenor"] = {
+},
+["Sfilatino - Draenor"] = {
+},
+["Cirus - Argent Dawn"] = {
+},
+["Pulcinhell - Draenor"] = {
+},
+["Panuozzo - Argent Dawn"] = {
+},
+["Panzarotto - Draenor"] = {
+},
+["Giandora - Argent Dawn"] = {
+},
+["Percoca - Draenor"] = {
+},
+["Branthar - Argent Dawn"] = {
+},
+["Nutellone - Draenor"] = {
+},
+["Ossosacro - Draenor"] = {
+},
+},
+}