+++ WTF/Account/NAME/SavedVariables/WaypointUI.lua
@@ -1,0 +1,9 @@
+
+WaypointDB_Global = {
+["WaypointDistanceTextType"] = 1,
+["lastLoadedVersion"] = 10203,
+["PrefFont"] = 1,
+["WaypointSystemType"] = 1,
+}
+WaypointDB_Global_Persistent = {
+}